Replace 'aspire start' with 'aspire run --detach'
- Add --detach/-d option to run command
- Move detached execution logic from StartCommand to RunCommand
- Remove StartCommand and its resources
- Update E2E test to use 'aspire run --detach'
- Stop command remains unchanged
