+++ src/Aspire.Cli/Commands/RootCommand.cs
@@ -24,7 +24,6 @@ internal sealed class RootCommand : BaseRootCommand
         NewCommand newCommand,
         InitCommand initCommand,
         RunCommand runCommand,
-        StartCommand startCommand,
         StopCommand stopCommand,
         AddCommand addCommand,
         PublishCommand publishCommand,
@@ -45,7 +44,6 @@ internal sealed class RootCommand : BaseRootCommand
         ArgumentNullException.ThrowIfNull(newCommand);
         ArgumentNullException.ThrowIfNull(initCommand);
         ArgumentNullException.ThrowIfNull(runCommand);
-        ArgumentNullException.ThrowIfNull(startCommand);
         ArgumentNullException.ThrowIfNull(stopCommand);
         ArgumentNullException.ThrowIfNull(addCommand);
         ArgumentNullException.ThrowIfNull(publishCommand);
@@ -112,7 +110,6 @@ internal sealed class RootCommand : BaseRootCommand
         Subcommands.Add(newCommand);
         Subcommands.Add(initCommand);
         Subcommands.Add(runCommand);
-        Subcommands.Add(startCommand);
         Subcommands.Add(stopCommand);
         Subcommands.Add(addCommand);
         Subcommands.Add(publishCommand);
