+++ src/Aspire.Cli/Commands/RunCommand.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
+using System.Diagnostics;
 using System.Globalization;
 using Aspire.Cli.Backchannel;
 using Aspire.Cli.Certificates;
@@ -36,6 +37,7 @@ internal sealed class RunCommand : BaseCommand
     private readonly TimeProvider _timeProvider;
     private readonly ILogger<RunCommand> _logger;
     private readonly IAppHostProjectFactory _projectFactory;
+    private readonly IAuxiliaryBackchannelMonitor _backchannelMonitor;
 
     public RunCommand(
         IDotNetCliRunner runner,
@@ -53,6 +55,7 @@ internal sealed class RunCommand : BaseCommand
         ICliHostEnvironment hostEnvironment,
         ILogger<RunCommand> logger,
         IAppHostProjectFactory projectFactory,
+        IAuxiliaryBackchannelMonitor backchannelMonitor,
         TimeProvider? timeProvider)
         : base("run", RunCommandStrings.Description, features, updateNotifier, executionContext, interactionService)
     {
@@ -67,6 +70,7 @@ internal sealed class RunCommand : BaseCommand
         ArgumentNullException.ThrowIfNull(hostEnvironment);
         ArgumentNullException.ThrowIfNull(logger);
         ArgumentNullException.ThrowIfNull(projectFactory);
+        ArgumentNullException.ThrowIfNull(backchannelMonitor);
 
         _runner = runner;
         _interactionService = interactionService;
@@ -81,11 +85,16 @@ internal sealed class RunCommand : BaseCommand
         _hostEnvironment = hostEnvironment;
         _logger = logger;
         _projectFactory = projectFactory;
+        _backchannelMonitor = backchannelMonitor;
         _timeProvider = timeProvider ?? TimeProvider.System;
 
         var projectOption = new Option<FileInfo?>("--project");
         projectOption.Description = RunCommandStrings.ProjectArgumentDescription;
         Options.Add(projectOption);
+
+        var detachOption = new Option<bool>("--detach", "-d");
+        detachOption.Description = RunCommandStrings.DetachArgumentDescription;
+        Options.Add(detachOption);
 
         if (ExtensionHelper.IsExtensionHost(InteractionService, out _, out _))
         {
@@ -100,11 +109,18 @@ internal sealed class RunCommand : BaseCommand
     protected override async Task<int> ExecuteAsync(ParseResult parseResult, CancellationToken cancellationToken)
     {
         var passedAppHostProjectFile = parseResult.GetValue<FileInfo?>("--project");
+        var detach = parseResult.GetValue<bool>("--detach");
         var isExtensionHost = ExtensionHelper.IsExtensionHost(InteractionService, out _, out _);
         var startDebugSession = isExtensionHost && parseResult.GetValue<bool>("--start-debug-session");
         var runningInstanceDetectionEnabled = _features.IsFeatureEnabled(KnownFeatures.RunningInstanceDetectionEnabled, defaultValue: true);
         // Force option kept for backward compatibility but no longer used since prompt was removed
         // var force = runningInstanceDetectionEnabled && parseResult.GetValue<bool>("--force");
+
+        // Handle detached mode - spawn child process and exit
+        if (detach)
+        {
+            return await ExecuteDetachedAsync(parseResult, passedAppHostProjectFile, cancellationToken);
+        }
 
         // A user may run `aspire run` in an Aspire terminal in VS Code. In this case, intercept and prompt
         // VS Code to start a debug session using the current directory
@@ -376,7 +392,7 @@ internal sealed class RunCommand : BaseCommand
         var appHostLabel = RunCommandStrings.AppHost;
         var dashboardLabel = RunCommandStrings.Dashboard;
         var logsLabel = RunCommandStrings.Logs;
-        var pidLabel = StartCommandStrings.ProcessId;
+        var pidLabel = RunCommandStrings.ProcessId;
 
         // Calculate column width based on all possible labels
         var labels = new List<string> { appHostLabel, dashboardLabel, logsLabel };
@@ -502,5 +518,237 @@ internal sealed class RunCommand : BaseCommand
 
             _resourceStates[resourceState.Resource] = resourceState;
         }
+    }
+
+    /// <summary>
+    /// Executes the run command in detached mode by spawning a child CLI process.
+    /// The parent waits for the auxiliary backchannel to become available, displays a summary, then exits
+    /// while the child continues running.
+    /// </summary>
+    /// <remarks>
+    /// <para><b>Failure Modes:</b></para>
+    /// <list type="number">
+    /// <item><b>Project not found</b>: No AppHost project found in the current directory or specified path.
+    /// Returns <see cref="ExitCodeConstants.FailedToFindProject"/>.</item>
+    /// <item><b>Failed to spawn child process</b>: Process.Start fails (e.g., executable not found).
+    /// Returns <see cref="ExitCodeConstants.FailedToDotnetRunAppHost"/>.</item>
+    /// <item><b>Child process exits early</b>: The child 'aspire run' process exits before the backchannel
+    /// is established (e.g., build failure, configuration error). Detected via WaitForExitAsync racing
+    /// with the poll delay. Shows exit code and log file path.
+    /// Returns <see cref="ExitCodeConstants.FailedToDotnetRunAppHost"/>.</item>
+    /// <item><b>Timeout waiting for backchannel</b>: The auxiliary backchannel socket doesn't appear
+    /// within 120 seconds. The child process is killed. Shows timeout message and log file path.
+    /// Returns <see cref="ExitCodeConstants.FailedToDotnetRunAppHost"/>.</item>
+    /// </list>
+    /// <para>On any failure, the log file path is displayed so the user can investigate.</para>
+    /// </remarks>
+    private async Task<int> ExecuteDetachedAsync(ParseResult parseResult, FileInfo? passedAppHostProjectFile, CancellationToken cancellationToken)
+    {
+        // Failure mode 1: Project not found
+        var searchResult = await _projectLocator.UseOrFindAppHostProjectFileAsync(
+            passedAppHostProjectFile,
+            MultipleAppHostProjectsFoundBehavior.Prompt,
+            createSettingsFile: false,
+            cancellationToken);
+
+        var effectiveAppHostFile = searchResult.SelectedProjectFile;
+
+        if (effectiveAppHostFile is null)
+        {
+            return ExitCodeConstants.FailedToFindProject;
+        }
+
+        _logger.LogDebug("Starting AppHost in background: {AppHostPath}", effectiveAppHostFile.FullName);
+
+        // Compute the expected auxiliary socket path hash for this AppHost
+        // The Connections dictionary is keyed by hash, not the full path
+        var expectedSocketPath = AppHostHelper.ComputeAuxiliarySocketPath(
+            effectiveAppHostFile.FullName,
+            ExecutionContext.HomeDirectory.FullName);
+        // We know the format is valid since we just computed it with ComputeAuxiliarySocketPath
+        var expectedHash = AppHostHelper.ExtractHashFromSocketPath(expectedSocketPath)!;
+
+        _logger.LogDebug("Waiting for socket: {SocketPath}, Hash: {Hash}", expectedSocketPath, expectedHash);
+
+        // Build the arguments for the child CLI process
+        var args = new List<string>
+        {
+            "run",
+            "--non-interactive",
+            "--project",
+            effectiveAppHostFile.FullName
+        };
+
+        // Pass through any unmatched tokens (but not --detach since child shouldn't detach again)
+        foreach (var token in parseResult.UnmatchedTokens)
+        {
+            if (token != "--detach" && token != "-d")
+            {
+                args.Add(token);
+            }
+        }
+
+        // Get the path to the current executable
+        var cliExecutable = Environment.ProcessPath ?? "aspire";
+
+        _logger.LogDebug("Spawning child CLI: {Executable} with args: {Args}", cliExecutable, string.Join(" ", args));
+        _interactionService.DisplayMessage("rocket", RunCommandStrings.StartingAppHostInBackground);
+
+        // Failure mode 2: Failed to spawn child process
+        var startInfo = new ProcessStartInfo
+        {
+            FileName = cliExecutable,
+            UseShellExecute = false,
+            CreateNoWindow = true,
+            RedirectStandardOutput = false,
+            RedirectStandardError = false,
+            RedirectStandardInput = false,
+            WorkingDirectory = effectiveAppHostFile.Directory?.FullName ?? ExecutionContext.WorkingDirectory.FullName
+        };
+
+        foreach (var arg in args)
+        {
+            startInfo.ArgumentList.Add(arg);
+        }
+
+        Process? childProcess;
+        try
+        {
+            childProcess = Process.Start(startInfo);
+            if (childProcess is null)
+            {
+                _interactionService.DisplayError(RunCommandStrings.FailedToStartAppHost);
+                return ExitCodeConstants.FailedToDotnetRunAppHost;
+            }
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, "Failed to start child CLI process");
+            _interactionService.DisplayError(RunCommandStrings.FailedToStartAppHost);
+            return ExitCodeConstants.FailedToDotnetRunAppHost;
+        }
+
+        _logger.LogDebug("Child CLI process started with PID: {PID}", childProcess.Id);
+
+        // Failure modes 3 & 4: Wait for the auxiliary backchannel to become available
+        // - Mode 3: Child exits early (build failure, config error, etc.)
+        // - Mode 4: Timeout waiting for backchannel (120 seconds)
+        AppHostAuxiliaryBackchannel? backchannel = null;
+        var startTime = _timeProvider.GetUtcNow();
+        var timeout = TimeSpan.FromSeconds(120);
+        var childExitedEarly = false;
+        var childExitCode = 0;
+
+        backchannel = await _interactionService.ShowStatusAsync(
+            RunCommandStrings.WaitingForAppHostToStart,
+            async () =>
+            {
+                while (_timeProvider.GetUtcNow() - startTime < timeout)
+                {
+                    cancellationToken.ThrowIfCancellationRequested();
+
+                    // Failure mode 3: Child process exited early
+                    if (childProcess.HasExited)
+                    {
+                        childExitedEarly = true;
+                        childExitCode = childProcess.ExitCode;
+                        _logger.LogWarning("Child CLI process exited with code {ExitCode}", childExitCode);
+                        return null;
+                    }
+
+                    // Trigger a scan and try to connect
+                    await _backchannelMonitor.ScanAsync(cancellationToken).ConfigureAwait(false);
+
+                    // Check if we can find a connection for this AppHost (keyed by hash)
+                    if (_backchannelMonitor.Connections.TryGetValue(expectedHash, out var connection))
+                    {
+                        return connection;
+                    }
+
+                    // Wait a bit before trying again, but short-circuit if the child process exits
+                    try
+                    {
+                        await childProcess.WaitForExitAsync(cancellationToken).WaitAsync(TimeSpan.FromMilliseconds(500), cancellationToken).ConfigureAwait(false);
+                        // If we get here, the process exited - we'll catch it at the top of the next iteration
+                    }
+                    catch (TimeoutException)
+                    {
+                        // Expected - the 500ms delay elapsed without the process exiting
+                    }
+                }
+
+                // Failure mode 4: Timeout - loop exited without finding connection
+                return null;
+            });
+
+        // Handle failure cases - show specific error and log file path
+        if (backchannel is null)
+        {
+            // Compute the expected log file path for error message
+            var expectedLogFile = AppHostHelper.GetLogFilePath(
+                childProcess.Id,
+                ExecutionContext.HomeDirectory.FullName,
+                _timeProvider);
+
+            if (childExitedEarly)
+            {
+                _interactionService.DisplayError(string.Format(
+                    CultureInfo.CurrentCulture,
+                    RunCommandStrings.AppHostExitedWithCode,
+                    childExitCode));
+            }
+            else
+            {
+                _interactionService.DisplayError(RunCommandStrings.TimeoutWaitingForAppHost);
+
+                // Try to kill the child process if it's still running (timeout case)
+                if (!childProcess.HasExited)
+                {
+                    try
+                    {
+                        childProcess.Kill();
+                    }
+                    catch
+                    {
+                        // Ignore errors when killing
+                    }
+                }
+            }
+
+            // Always show log file path for troubleshooting
+            _interactionService.DisplayMessage("magnifying_glass_tilted_right", string.Format(
+                CultureInfo.CurrentCulture,
+                RunCommandStrings.CheckLogsForDetails,
+                expectedLogFile.FullName));
+
+            return ExitCodeConstants.FailedToDotnetRunAppHost;
+        }
+
+        var appHostInfo = backchannel.AppHostInfo;
+
+        // Get the dashboard URLs
+        var dashboardUrls = await backchannel.GetDashboardUrlsAsync(cancellationToken).ConfigureAwait(false);
+
+        // Get the log file path
+        var logFile = AppHostHelper.GetLogFilePath(
+            appHostInfo?.ProcessId ?? childProcess.Id,
+            ExecutionContext.HomeDirectory.FullName,
+            _timeProvider);
+
+        // Display success UX using shared rendering
+        var appHostRelativePath = Path.GetRelativePath(ExecutionContext.WorkingDirectory.FullName, effectiveAppHostFile.FullName);
+        var pid = appHostInfo?.ProcessId ?? childProcess.Id;
+        RenderAppHostSummary(
+            _ansiConsole,
+            appHostRelativePath,
+            dashboardUrls?.BaseUrlWithLoginToken,
+            codespacesUrl: null,
+            logFile.FullName,
+            pid);
+        _ansiConsole.WriteLine();
+
+        _interactionService.DisplaySuccess(RunCommandStrings.AppHostStartedSuccessfully);
+
+        return ExitCodeConstants.Success;
     }
 }
