+++ src/Aspire.Cli/Resources/RunCommandStrings.resx
@@ -126,6 +126,9 @@
   <data name="ProjectArgumentDescription" xml:space="preserve">
     <value>The path to the Aspire AppHost project file.</value>
   </data>
+  <data name="DetachArgumentDescription" xml:space="preserve">
+    <value>Run the AppHost in the background and exit after it starts.</value>
+  </data>
   <data name="WatchArgumentDescription" xml:space="preserve">
     <value>Start project resources in watch mode.</value>
   </data>
@@ -201,5 +204,29 @@
   </data>
   <data name="RunningInstanceStopped" xml:space="preserve">
     <value>Running instance stopped successfully.</value>
+  </data>
+  <data name="StartingAppHostInBackground" xml:space="preserve">
+    <value>Starting Aspire apphost in the background...</value>
+  </data>
+  <data name="AppHostStartedSuccessfully" xml:space="preserve">
+    <value>AppHost started successfully.</value>
+  </data>
+  <data name="FailedToStartAppHost" xml:space="preserve">
+    <value>Failed to start the AppHost.</value>
+  </data>
+  <data name="WaitingForAppHostToStart" xml:space="preserve">
+    <value>Waiting for apphost to start...</value>
+  </data>
+  <data name="AppHostExitedWithCode" xml:space="preserve">
+    <value>AppHost process exited with code {0}.</value>
+  </data>
+  <data name="TimeoutWaitingForAppHost" xml:space="preserve">
+    <value>Timeout waiting for AppHost to start.</value>
+  </data>
+  <data name="CheckLogsForDetails" xml:space="preserve">
+    <value>Check logs for details: {0}</value>
+  </data>
+  <data name="ProcessId" xml:space="preserve">
+    <value>PID</value>
   </data>
 </root>
