+++ tests/Aspire.Cli.EndToEnd.Tests/StartStopTests.cs
@@ -98,8 +98,8 @@ public sealed class StartStopTests(ITestOutputHelper output)
             .Enter()
             .WaitForSuccessPrompt(counter);
 
-        // Start the AppHost in the background using aspire start
-        sequenceBuilder.Type("aspire start")
+        // Start the AppHost in the background using aspire run --detach
+        sequenceBuilder.Type("aspire run --detach")
             .Enter()
             .WaitUntil(s => waitForAppHostStartedSuccessfully.Search(s).Count > 0, TimeSpan.FromMinutes(3))
             .WaitForSuccessPrompt(counter);
